+++ android/src/main/java/ai/quickpose/reactnative/QuickPosePackage.kt
@@ -7,7 +7,7 @@ import com.facebook.react.uimanager.ViewManager
 
 class QuickPosePackage : ReactPackage {
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
-        return emptyList()
+        return listOf(QuickPoseCaptureModule(reactContext))
     }
 
     override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {